Changed ProjectView and TeamListView confusion with naming conventions and routing
